third_party/blink/web_tests/fast/events/touch/gesture/gesture-tap-scrolled-expected.txt

Verifies that a tap occurring in a scrolled page has the correct co-ordinates

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending GestureTapDown to 50,50
Sending GestureShowPress
Sending GestureTap
Received mousemove in page
PASS event.screenX is point.x
PASS event.screenY is point.y
PASS event.clientX is point.x
PASS event.clientY is point.y
PASS event.pageX is point.x + scrollOffset.x
PASS event.pageY is point.y + scrollOffset.y
Received mousedown in page
PASS event.screenX is point.x
PASS event.screenY is point.y
PASS event.clientX is point.x
PASS event.clientY is point.y
PASS event.pageX is point.x + scrollOffset.x
PASS event.pageY is point.y + scrollOffset.y
Received mouseup in page
PASS event.screenX is point.x
PASS event.screenY is point.y
PASS event.clientX is point.x
PASS event.clientY is point.y
PASS event.pageX is point.x + scrollOffset.x
PASS event.pageY is point.y + scrollOffset.y
Received click in page
PASS event.screenX is point.x
PASS event.screenY is point.y
PASS event.clientX is point.x
PASS event.clientY is point.y
PASS event.pageX is point.x + scrollOffset.x
PASS event.pageY is point.y + scrollOffset.y
PASS eventReceived.mousemove is true
PASS eventReceived.mousedown is true
PASS eventReceived.mouseup is true
PASS eventReceived.click is true
PASS successfullyParsed is true

TEST COMPLETE
